+++ packages/iov-tendermint-rpc/src/index.ts
@@ -1,5 +1,5 @@
 // exported to access version-specific hashing
-export { v0_27 } from "./v0-27";
+export { v0_29 } from "./v0-29";
 export { v0_31 } from "./v0-31";
 
 export { Client } from "./client";