+++ packages/iov-tendermint-rpc/types/v0-29/index.d.ts
@@ -1,2 +1,2 @@
 import { Adaptor } from "../adaptor";
-export declare const v0_27: Adaptor;
+export declare const v0_29: Adaptor;