+++ scripts/tendermint/all_start.sh
@@ -4,7 +4,6 @@ command -v shellcheck > /dev/null && shellcheck "$0"
 
 # Find latest patch releases at https://hub.docker.com/r/tendermint/tendermint/tags/
 declare -a TM_VERSIONS
-TM_VERSIONS[27]=0.27.4
 TM_VERSIONS[29]=0.29.2
 TM_VERSIONS[30]=v0.30.4
 TM_VERSIONS[31]=v0.31.5