+++ scripts/tendermint/all_stop.sh
@@ -3,7 +3,6 @@ set -o errexit -o nounset -o pipefail
 command -v shellcheck > /dev/null && shellcheck "$0"
 
 declare -a TM_VERSIONS
-TM_VERSIONS[27]=0.27.4
 TM_VERSIONS[29]=0.29.2
 TM_VERSIONS[30]=v0.30.4
 TM_VERSIONS[31]=v0.31.5