+++ CHANGELOG.md
@@ -67,7 +67,9 @@ Breaking changes
   `JsonRpcErrorResponse` is now of type `JsonRpcId`, which can be a string or a
   number.
 - @iov/jsonrpc: Remove `parseJsonRpcResponse` and `parseJsonRpcError`.
-- @iov/tendermint-rpc: Remove support for Tendermint 0.25.x.
+- @iov/tendermint-rpc: Remove support for Tendermint 0.25.x – 0.28.x.
+- @iov/tendermint-rpc: Rename `v0_27` to `v0_29`, which is now the adaptor for
+  0.29.x and 0.30.x.
 - @iov/tendermint-rpc: Rename `ConsensusParams.blockSize` to `.block` and
   `BlockSizeParams` to `BlockParams`.